requirements.txt
tensorflow==2.15.0
numpy==1.26.4
pandas==2.3.3
Pillow==10.4.0
streamlit==1.32.0
streamlit-option-menu==0.3.12